Refactor Admin Dashboard: Split components into separate files and remove unused code
- Created AdminOverviewTab component to handle the overview tab content.
- Created AdminCategoriesTab component to manage the categories tab content.
- Created AdminTeamsTab component to manage the teams tab content.
- Removed inline category and team management logic from AdminDashboardClient.
- Updated Badge component to use span instead of div for better semantics.
- Cleaned up unused state and mutation logic related to category management in AdminDashboardClient.
