+++ src/app/dashboard/admin/components/AdminCategoriesTab.tsx
@@ -1,0 +1,61 @@
+import { Badge } from "@/presentation/components/ui/badge";
+import { Button } from "@/presentation/components/ui/button";
+import { Card, CardContent } from "@/presentation/components/ui/card";
+import {
+  Table,
+  TableBody,
+  TableCell,
+  TableHead,
+  TableHeader,
+  TableRow,
+} from "@/presentation/components/ui/table";
+
+interface CategoryListItem {
+  id: string;
+  name: string;
+  type: string;
+}
+
+interface AdminCategoriesTabProps {
+  categories: CategoryListItem[];
+}
+
+export function AdminCategoriesTab({ categories }: AdminCategoriesTabProps) {
+  return (
+    <div className="space-y-4">
+      <div className="flex items-center justify-between">
+        <h3 className="text-lg font-semibold">Categorias</h3>
+      </div>
+      <Card>
+        <CardContent className="p-0">
+          <Table>
+            <TableHeader>
+              <TableRow>
+                <TableHead>Nome</TableHead>
+                <TableHead>Tipo</TableHead>
+                <TableHead>Ações</TableHead>
+              </TableRow>
+            </TableHeader>
+            <TableBody>
+              {categories.map((category) => (
+                <TableRow key={category.id}>
+                  <TableCell className="font-medium">{category.name}</TableCell>
+                  <TableCell>
+                    <Badge variant={category.type === "RESCUE" ? "default" : "secondary"}>
+                      {category.type === "RESCUE" ? "Resgate" : "Artística"}
+                    </Badge>
+                  </TableCell>
+                  <TableCell>
+                    <Button variant="outline" size="sm" asChild>
+                      <a href={`/ranking?categoryId=${category.id}`}>Ver Ranking</a>
+                    </Button>
+                  </TableCell>
+                </TableRow>
+              ))}
+            </TableBody>
+          </Table>
+        </CardContent>
+      </Card>
+    </div>
+  );
+}
